figma-transformer: extract sticky layout coordinator
